Rebase to use CI-RC03. Moved InsufficientBalanceException into the API package for token selection. Fixed bug in evolvable token flows where sigs where checked before they were all added.
