+++ workflows/src/main/kotlin/com/r3/corda/lib/tokens/workflows/internal/flows/confidential/AnonymisePartiesFlowHandler.kt
@@ -1,8 +1,7 @@
 package com.r3.corda.lib.tokens.workflows.internal.flows.confidential
 
 import co.paralleluniverse.fibers.Suspendable
-import com.r3.corda.lib.ci.ProvideKeyFlow
-import com.r3.corda.lib.ci.RequestKeyResponder
+import com.r3.corda.lib.ci.workflows.ProvideKeyFlow
 import net.corda.core.flows.FlowLogic
 import net.corda.core.flows.FlowSession
 import net.corda.core.utilities.unwrap
